CausalGUI/causalproject/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.7.0
pandas==2.2.3
networkx==3.4.2
matplotlib==3.10.1
dowhy==0.14
numpy>=2.2,<3
openai>=1.66,<2
